+++ app/build.gradle
@@ -26,17 +26,14 @@ repositories {
 dependencies {
     def room_version = '2.1.0'
 
-    implementation 'com.google.code.gson:gson:2.8.2'
-
-    implementation "androidx.room:room-runtime:$room_version"
-    implementation 'androidx.preference:preference:1.1.0-rc01'
+	implementation "androidx.room:room-runtime:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.1.0-rc01'
     implementation 'com.google.android.material:material:1.1.0-alpha07'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta2'
 
-    //implementation 'androidx.recyclerview:recyclerview:1.0.0-beta01'
+	//implementation 'androidx.recyclerview:recyclerview:1.0.0-beta01'
     implementation 'androidx.vectordrawable:vectordrawable:1.1.0-rc01'
 
     //graphics
