+++ app/src/main/AndroidManifest.xml
@@ -10,16 +10,11 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".SettingsActivity"
-            android:label="@string/title_activity_settings"></activity>
-        <activity android:name=".Calendar_Activity">
-
-            <!--
-            android:parentActivityName=".MainActivity">
+            android:name=".Calendar_Activity">
+            <!--android:parentActivityName=".MainActivity">
             <meta-data
                 android:name="android.calendar.PARENT_ACTIVITY"
-                android:value=".MainActivity"/>
-            -->
+                android:value=".MainActivity"/>-->
         </activity>
         <activity
             android:name=".MainActivity"
@@ -31,12 +26,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AddFoodActivity"
-            android:parentActivityName=".FoodListActivity"></activity>
+            android:name= ".AddFoodActivity"
+            android:parentActivityName=".FoodListActivity">
+        </activity>
         <activity
             android:name=".FoodListActivity"
             android:label="@string/title_activity_food_list"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
+            android:theme="@style/AppTheme.NoActionBar">
+        </activity>
         <activity
             android:name=".DetailActivity"
             android:parentActivityName=".Calendar_Activity">
@@ -44,17 +41,7 @@
                 android:name="android.detail.PARENT_ACTIVITY"
                 android:value=".Calendar_Activity" />
         </activity>
-        <activity
-            android:name=".AddFoodPlan">
-        </activity>
-        <activity
-            android:name=".FoodPlanActivity"
-            android:label="FoodPlan"
-            android:theme="@style/AppTheme.NoActionBar">
-        </activity>
-        <activity
-            android:name=".Food_Main">
-        </activity>
+        <activity android:name=".SettingsActivity"></activity>
     </application>
 
 </manifest>
