+++ app/src/main/java/com/umn/appdev/fitu/FoodActivity.java
@@ -9,14 +9,14 @@ import android.view.MenuItem;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.umn.appdev.fitu.R;
-import com.umn.appdev.fitu.utils.BottomNavigationViewHelper;
+import com.umn.appdev.fitu.Utils.BottomNavigationViewHelper;
 
 import androidx.appcompat.app.AppCompatActivity;
-//what is this class for?
+
 public class FoodActivity extends AppCompatActivity {
 
-    //private static final String TAG = "FoodActivity";
-    //private static final int ACTIVITY_NUM = 2;
+    private static final String TAG = "FoodActivity";
+    private static final int ACTIVITY_NUM = 2;
 
     private Context mContext = FoodActivity.this;
 
@@ -25,9 +25,9 @@ public class FoodActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        //setupBottomNavigationView();
+        setupBottomNavigationView();
     }
-    /**
+
     private void setupBottomNavigationView(){
         Log.d(TAG, "setupBottomNavigationView: setting up BottomNavigationView");
         BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.navigation);
@@ -36,5 +36,6 @@ public class FoodActivity extends AppCompatActivity {
         MenuItem menuItem = menu.getItem(ACTIVITY_NUM);
         menuItem.setChecked(true);
     }
-    **/
+
+
 }
