+++ app/src/main/res/layout/activity_add_food.xml
@@ -23,7 +23,6 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="165dp"
         android:layout_marginEnd="55dp"
-        android:layout_marginRight="55dp"
         android:ems="10"
         android:inputType="number"
         app:layout_constraintEnd_toEndOf="parent"
@@ -36,7 +35,6 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
         android:layout_marginEnd="60dp"
-        android:layout_marginRight="60dp"
         android:ems="10"
         android:inputType="number"
         app:layout_constraintEnd_toEndOf="parent"
@@ -61,10 +59,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="32dp"
-        android:layout_marginLeft="32dp"
         android:layout_marginTop="52dp"
         android:layout_marginEnd="72dp"
-        android:layout_marginRight="72dp"
         android:text="Proteins"
         app:layout_constraintEnd_toStartOf="@+id/protein_input"
         app:layout_constraintStart_toStartOf="parent"
@@ -114,7 +110,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginEnd="26dp"
-        android:layout_marginRight="26dp"
         android:layout_marginBottom="34dp"
         android:ems="10"
         android:inputType="textPersonName"
