+++ app/src/main/res/layout/activity_food_list.xml
@@ -23,7 +23,7 @@
         app:srcCompat="@android:drawable/ic_dialog_email" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
-    <!--
+
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/navigation"
         android:layout_width="0dp"
@@ -35,5 +35,5 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:menu="@menu/navigation" />
-        -->
+
 </androidx.constraintlayout.widget.ConstraintLayout>
