--- app/src/main/res/menu/settings_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/food_settings"
-        android:orderInCategory="100"
-        android:title="settings"
-        app:showAsAction="never" />
-</menu>
