--- app/src/main/res/xml/root_preferences.xml
@@ -1,43 +1,0 @@
-<!--
-  ~ Copyright 2018 The app Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<androidx.preference.PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <PreferenceCategory app:title="macronutrients">
-
-        <EditTextPreference
-            app:key="proteins"
-            android:inputType="number"
-            app:title="proteins"
-            app:useSimpleSummaryProvider="true" />
-        <EditTextPreference
-            app:key="carbs"
-            android:inputType="number"
-            app:title="carbs"
-            app:useSimpleSummaryProvider="true" />
-        <EditTextPreference
-            app:key="fats"
-            android:inputType="number"
-            app:title="fats"
-            app:useSimpleSummaryProvider="true" />
-
-
-    </PreferenceCategory>
-
-
-
-</androidx.preference.PreferenceScreen>
